+++ backend/db/models/user.model.js
@@ -15,6 +15,7 @@ const userSchema = mongoose.Schema({
   },
   email: {
     type: String,
+    required: true,
   },
 });
