+++ backend/routes/user.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const User = require('../models/user.model');
+const User = require('../db/models/user.model');
 
 const app = express();
